ia-data-analyzer/mcp_server/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0
pyyaml==6.0.1
python-multipart==0.0.6

